브랜디 메인: 컴포넌트 쪼개기
